+++ modules/businessProcess/ui/src/sap/commerce/toolset/businessProcess/options/BpProjectConfigurableProvider.kt
@@ -16,34 +16,52 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.acl.options
+package sap.commerce.toolset.businessProcess.options
 
 import com.intellij.openapi.options.BoundSearchableConfigurable
 import com.intellij.openapi.options.ConfigurableProvider
 import com.intellij.openapi.project.Project
 import com.intellij.ui.dsl.builder.bindSelected
 import com.intellij.ui.dsl.builder.panel
+import com.intellij.ui.layout.selected
+import sap.commerce.toolset.i18n
 import sap.commerce.toolset.isHybrisProject
 import sap.commerce.toolset.settings.yDeveloperSettings
+import javax.swing.JCheckBox
 
-class ProjectAclSettingsConfigurableProvider(private val project: Project) : ConfigurableProvider() {
+class BpProjectConfigurableProvider(val project: Project) : ConfigurableProvider() {
 
     override fun canCreateConfigurable() = project.isHybrisProject
     override fun createConfigurable() = SettingsConfigurable(project)
 
-    class SettingsConfigurable(private val project: Project) : BoundSearchableConfigurable(
-        "Access Control Lists", "hybris.acl.settings"
+    class SettingsConfigurable(project: Project) : BoundSearchableConfigurable(
+        i18n("hybris.settings.project.bp.title"), "[y] SAP CX Business Process configuration."
     ) {
 
         private val developerSettings = project.yDeveloperSettings
-        private val mutableSettings = developerSettings.aclSettings.mutable()
+        private val mutableSettings = developerSettings.bpSettings.mutable()
+
+        private lateinit var foldingEnableCheckBox: JCheckBox
 
         override fun createPanel() = panel {
             group("Code Folding") {
                 row {
-                    checkBox("Enable code folding")
+                    foldingEnableCheckBox = checkBox("Enable code folding")
                         .bindSelected(mutableSettings.folding::enabled)
                         .component
+                }
+                group("Table-Like Folding", true) {
+                    row {
+                        checkBox("Action transitions")
+                            .bindSelected(mutableSettings.folding::tablifyActionTransitions)
+                            .enabledIf(foldingEnableCheckBox.selected)
+                        checkBox("Case choices")
+                            .bindSelected(mutableSettings.folding::tablifyCaseChoices)
+                            .enabledIf(foldingEnableCheckBox.selected)
+                        checkBox("Ends")
+                            .bindSelected(mutableSettings.folding::tablifyEnds)
+                            .enabledIf(foldingEnableCheckBox.selected)
+                    }
                 }
             }
         }
@@ -51,7 +69,7 @@ class ProjectAclSettingsConfigurableProvider(private val project: Project) : Con
         override fun apply() {
             super.apply()
 
-            developerSettings.aclSettings = mutableSettings.immutable()
+            developerSettings.bpSettings = mutableSettings.immutable()
         }
     }
 }
