+++ modules/ccv2/ui/resources/META-INF/sap.commerce.toolset-ccv2-ui.xml
@@ -23,7 +23,7 @@
         <projectConfigurable id="sap.commerce.toolset.ccv2.exec.settings" parentId="hybris.project.integrations.settings"
                              nonDefaultProject="true" dynamic="true"
                              displayName="CCv2" order="first"
-                             provider="sap.commerce.toolset.ccv2.options.ProjectCCv2ExecSettingsConfigurableProvider"/>
+                             provider="sap.commerce.toolset.ccv2.options.CCv2ExecProjectSettingsConfigurableProvider"/>
     </extensions>
 
     <actions>
