+++ modules/flexibleSearch/ui/resources/META-INF/sap.commerce.toolset-flexibleSearch-ui.xml
@@ -25,7 +25,7 @@
                              nonDefaultProject="true" dynamic="true"
                              bundle="i18n.HybrisBundle" key="hybris.settings.project.fxs.title"
                              order="after hybris.project.integrations.settings"
-                             provider="sap.commerce.toolset.flexibleSearch.options.ProjectFlexibleSearchSettingsConfigurableProvider"/>
+                             provider="sap.commerce.toolset.flexibleSearch.options.FlexibleSearchProjectSettingsConfigurableProvider"/>
     </extensions>
 
     <actions>
