+++ modules/flexibleSearch/ui/src/sap/commerce/toolset/flexibleSearch/actionSystem/FlexibleSearchOpenSettingsAction.kt
@@ -19,11 +19,11 @@
 package sap.commerce.toolset.flexibleSearch.actionSystem
 
 import sap.commerce.toolset.actionSystem.OpenSettingsAction
-import sap.commerce.toolset.flexibleSearch.options.ProjectFlexibleSearchSettingsConfigurableProvider
+import sap.commerce.toolset.flexibleSearch.options.FlexibleSearchProjectSettingsConfigurableProvider
 import sap.commerce.toolset.i18n
 
 class FlexibleSearchOpenSettingsAction : OpenSettingsAction(
-    configurableClass = ProjectFlexibleSearchSettingsConfigurableProvider.SettingsConfigurable::class.java,
+    configurableClass = FlexibleSearchProjectSettingsConfigurableProvider.SettingsConfigurable::class.java,
     text = i18n("hybris.fxs.actions.open_settings"),
     description = i18n("hybris.fxs.actions.open_settings.description"),
 )
