+++ modules/flexibleSearch/ui/src/sap/commerce/toolset/flexibleSearch/options/FlexibleSearchProjectSettingsConfigurableProvider.kt
@@ -34,7 +34,7 @@ import sap.commerce.toolset.settings.state.ReservedWordsCase
 import sap.commerce.toolset.settings.yDeveloperSettings
 import javax.swing.JCheckBox
 
-class ProjectFlexibleSearchSettingsConfigurableProvider(private val project: Project) : ConfigurableProvider() {
+class FlexibleSearchProjectSettingsConfigurableProvider(private val project: Project) : ConfigurableProvider() {
 
     override fun canCreateConfigurable() = project.isHybrisProject
     override fun createConfigurable() = SettingsConfigurable(project)
