+++ modules/polyglotQuery/ui/resources/META-INF/sap.commerce.toolset-polyglotQuery-ui.xml
@@ -28,7 +28,7 @@
         <projectConfigurable id="hybris.project.pgq.settings" parentId="hybris.project.settings" nonDefaultProject="true" dynamic="true"
                              bundle="i18n.HybrisBundle" key="hybris.settings.project.pgq.title"
                              order="after hybris.project.fxs.settings"
-                             provider="sap.commerce.toolset.polyglotQuery.options.ProjectPolyglotQuerySettingsConfigurableProvider"/>
+                             provider="sap.commerce.toolset.polyglotQuery.options.PolyglotQueryProjectSettingsConfigurableProvider"/>
     </extensions>
 
     <actions>
