+++ modules/solr/ui/resources/META-INF/sap.commerce.toolset-solr-ui.xml
@@ -22,7 +22,7 @@
         <projectConfigurable id="sap.commerce.toolset.solr.exec.settings" parentId="hybris.project.integrations.settings"
                              nonDefaultProject="true" dynamic="true"
                              displayName="Solr" order="first"
-                             provider="sap.commerce.toolset.solr.options.ProjectSolrExecSettingsConfigurableProvider"/>
+                             provider="sap.commerce.toolset.solr.options.SolrExecProjectSettingsConfigurableProvider"/>
     </extensions>
 
     <actions>
