+++ resources/META-INF/plugin-internal.xml
@@ -33,7 +33,7 @@
         <projectConfigurable id="hybris.project.integrations.settings" parentId="hybris.project.settings"
                              nonDefaultProject="true" dynamic="true"
                              displayName="Integrations" order="first"
-                             provider="sap.commerce.toolset.options.ProjectIntegrationsSettingsConfigurableProvider"/>
+                             provider="sap.commerce.toolset.options.IntegrationsProjectSettingsConfigurableProvider"/>
     </extensions>
 
     <extensions defaultExtensionNs="com.intellij">
